Encapsulated JS within an anonymous function
This should help prevent namespace conflicts and allow you to use the $
shortcut. The only potential issue is that functions and variables
which once were global are no longer accessible within the global
scope. If other areas of the plugin depend on them or other plugins are
using them, they will break.