Update the Registration logic. Per #287
This does two main things. 1. It makes it NOT check the domain every time it checks if it's registered. If there's a premium code in the database, just roll with it. 2. It now ONLY checks the domain during the monthly registration check. During this time, if it doesn't match, it will automatically migrate itself onto the new domain that it's detecting.