Updated the raw shares fix.
Instead of having the user opt into the fix, I'll hvae them opt out. It now ONLY adds the filter to the content filter if it's on single.php view or if the user's theme uses full content rather than excerpts sitewide (which is rare) in which case they can toggle a button on to have the filter added everywhere else also.