Set up 2 series of meta tag hook filters.
I've created two sets of meta tag hook filters. 1 is for creating the values and will store them in the ['meta_tag_values'] array. The second hook filter will then take the assigned values and compile them into html for output. This will allow more efficient processing of the values by us as well as allowing third-parties to be able to reach in and modify very specific values prior to them being turned into HTML for output. The next step is to create the functions that will tie into these hooks.