The open graph values are now returned as values rather than as html.
The first filter hook function for returning the values is now complete. We still need to create the values for Twitter cards and then create the filter hook functions that compile them into html. Additionally, for all of this to work properly, I'm going to need to refactor the cache rebuild process to ensure all the values are stored properly and made available to these new functions.