-Added timestamp output from server -Added JS browser timestamp comparison -Made ajax call for swp_cache_trigger only work if within timelimt -Tested numerous times both above and below the time limit. -Made Facebook check (socialWarfarePlugin.fetchShares) only work if within timelimit --Had to move fetchShares call to be within document.onready because it needs to know if the time is within limits
FACEBOOK FETCH TEST INADEQUACY
-The fetchShares function did not output a message indicating a Facebook
response.  This could be for a number of reasons, namely the
test environment is not yet configured with a Facebook account, and the
test site/urls have 0 shares.  I recommend testing this change
on a page that has Facebook shares.

TIME DIFFERENCE PROBLEM
See previous GitHub message:
https://github.com/warfare-plugins/social-warfare-pro/issues/12