Added additional info about submitting issues
Attempting to streamline the process of user's who want to submit issues to the GitHub tracker and make this tracker work more seamlessly with the support documentation and the email support team. If we can get all 3 working in harmony, we should be able to solve many, many more issues for folks.