Updated the include reference for the registration template
This commit updates the "require_once" reference that pulls the
registration template from the pro plugin and adds the registration form
to the admin options page. This is because I reorganized the location of
all the files in both the free version and the pro version of the plugin
to make it more organized and logical in their structure.