Updating the condition fields function. #463
I've got the function added to admin.js so it loads everywhere in the admin side of the site, and the function is now in the global scope so it can be called from anywhere. However, I can't seem to get updated values from the fields on the widget form. Once we solve that, this task will be done.