Deferred rendering and passing parameters (see notes)
Because we need addons to be able to hook into the options page, we can not initialize the registration tab in the constructor. Instead, this will be the only tab which initializes at render-time. Right before init_registration(), we fetch all the applied addons so we can pass them to init_registration( $addons ).