Switched registration from data to regular attribute.
I wanted to use the data-{name} attribute rather than just creating a random attribute because this makes the data accessible as a dataset in memory. However, when manipulating this dataset, we are not actually changing the content of the DOM. We need to change the attribute in the DOM for CSS to take effect. Therefore, these attributes have been returned to their original state (not as data).
So
data-swp-registrations => swp-registrations