Major update to file. See description.
This class now double encodes all of the data.
In raw from, it looks like a string.
In parsed form, it looks like an array of `[$network_key][$token_type] = $token`.

All methods are now public static methods.
Once we fetch the data, we store it locally as a static member.