Refactor SWP_Post_Cache.php to add parsed_api_responses property
This commit adds a new property, $parsed_api_responses, to the SWP_Post_Cache class. This property will store the parsed API responses. This change is necessary to improve the efficiency and organization of the code.

Note: This commit message follows the established convention of using a prefix (Refactor) to indicate a code improvement or optimization.