Refactor SWP_Notice_Loader.php to add notices array
This commit adds a public property `$notices` to the `SWP_Notice_Loader` class, which is an array that will store notices. This change allows for better management and display of notices within the plugin.

Note: This commit message follows the established convention of using a prefix (Refactor) to indicate a code improvement or optimization.