Refactor SWP_Notice_Loader.php to load persistent notices
This commit adds a new method `load_persistent_notices()` to the `SWP_Notice_Loader` class. This method is responsible for loading the persistent notices from the database. The purpose of this change is to ensure that the persistent notices are properly loaded and displayed to users.

Note: This commit message follows the established convention of using a prefix (Refactor) to indicate a code improvement or optimization.