Refactor SWP_Notice.php to add notice_keys property
This commit adds a new public static property `$notice_keys` to the `SWP_Notice` class. This property will store the notice keys that have been registered. The purpose of this change is to improve the management and retrieval of notice keys within the plugin.

Note: This commit message follows the established convention of using a prefix (Refactor) to indicate a code improvement or optimization.