Refactor composer.json and composer.lock to update dependencies
This commit updates the `composer.json` and `composer.lock` files to modify the version constraint for the `10up/phpcs-composer` package. The version constraint is changed from `^3.0` to `3.0` in order to use the exact version specified. Additionally, the `content-hash` value in the `composer.lock` file is updated.

Note: This commit message follows the established convention of using a prefix (Refactor) to indicate a code improvement or optimization.